Revision del codigo

+++ BackEnd-SpringBoot/src/main/java/com/portfolio/backend/security/repository/UserRepository.java
@@ -1,14 +1,12 @@
 package com.portfolio.backend.security.repository;
 
-import com.portfolio.backend.security.models.User;
+import com.portfolio.backend.security.model.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long>{
+public interface UserRepository extends JpaRepository<User, Integer>{
     Optional<User> findByUserName(String userName);
-    
     boolean existsByUserName(String userName);
-
 }
